Refactor core package; add KeytoolCommandFactory to encapsulate keytool command options.
