+++ src/main/kotlin/edu/adarko22/jdkcerts/cli/command/jdk/InstallCertCliCommand.kt
@@ -7,8 +7,8 @@ import edu.adarko22.jdkcerts.cli.command.certPathOption
 import edu.adarko22.jdkcerts.cli.command.customJdkDirsOption
 import edu.adarko22.jdkcerts.cli.command.dryRunOption
 import edu.adarko22.jdkcerts.cli.command.keystorePasswordOption
-import edu.adarko22.jdkcerts.core.jdk.KeytoolCommand
-import edu.adarko22.jdkcerts.core.jdk.usecase.ExecuteKeytoolCommandUseCase
+import edu.adarko22.jdkcerts.core.jdk.keytool.model.KeytoolCommandFactory
+import edu.adarko22.jdkcerts.core.jdk.keytool.usecase.ExecuteKeytoolCommandUseCase
 import java.nio.file.Path
 
 /**
@@ -27,21 +27,7 @@ class InstallCertCliCommand(
     override fun help(context: Context) = "Install certificate across all JDK keystores"
 
     override fun run() {
-        val command =
-            KeytoolCommand
-                .Builder()
-                .addArg("-importcert")
-                .addArg("-noprompt")
-                .addArg("-trustcacerts")
-                .addArg("-alias")
-                .addArg(alias)
-                .addArg("-file")
-                .addArg("$certPath")
-                .addArg("-storepass")
-                .addArg(keystorePassword)
-                .withKeystoreResolution()
-                .build()
-
+        val command = KeytoolCommandFactory.installCertificateKeytoolCommand(alias, keystorePassword, certPath.toAbsolutePath().toString())
         val results = executeKeytoolCommandUseCase.execute(keytoolCommand = command, customJdkDirs, dryRun)
         executeKeytoolCommandCliPresenter.present(results, dryRun)
     }
