+++ src/main/kotlin/edu/adarko22/jdkcerts/core/jdk/DiscoverJdksUseCase.kt
@@ -1,6 +1,6 @@
-package edu.adarko22.jdkcerts.core.jdk.usecase
+package edu.adarko22.jdkcerts.core.jdk
 
-import edu.adarko22.jdkcerts.core.jdk.Jdk
+import edu.adarko22.jdkcerts.core.jdk.java.usecase.ResolveJavaInfoUseCase
 import edu.adarko22.jdkcerts.infra.system.JdkPathsDiscovery
 import edu.adarko22.jdkcerts.infra.system.KeystoreInfoResolver
 import java.nio.file.Path
