+++ src/main/kotlin/edu/adarko22/jdkcerts/core/jdk/java/usecase/ResolveJavaInfoUseCase.kt
@@ -1,8 +1,8 @@
-package edu.adarko22.jdkcerts.core.jdk.usecase
+package edu.adarko22.jdkcerts.core.jdk.java.usecase
 
 import edu.adarko22.jdkcerts.core.execution.ProcessRunner
-import edu.adarko22.jdkcerts.core.jdk.JavaInfo
-import edu.adarko22.jdkcerts.core.jdk.parser.JavaInfoParser
+import edu.adarko22.jdkcerts.core.jdk.java.model.JavaInfo
+import edu.adarko22.jdkcerts.core.jdk.java.parser.JavaInfoParser
 import java.nio.file.Path
 import kotlin.io.path.absolutePathString
