+++ src/main/kotlin/edu/adarko22/jdkcerts/infra/system/KeystoreInfoResolver.kt
@@ -1,7 +1,7 @@
 package edu.adarko22.jdkcerts.infra.system
 
-import edu.adarko22.jdkcerts.core.jdk.JavaInfo
-import edu.adarko22.jdkcerts.core.jdk.KeystoreInfo
+import edu.adarko22.jdkcerts.core.jdk.java.model.JavaInfo
+import edu.adarko22.jdkcerts.core.jdk.keytool.model.KeystoreInfo
 import java.nio.file.Path
 
 /**
